requirements.txt
prometheus-client==0.12.0
PyYAML==5.3.1
lxml==4.5.0
urllib3==1.25.8
